Add database, some fixes

+++ src/main/java/ru/practicum/shareit/item/controller/ItemController.java
@@ -15,23 +15,23 @@ public class ItemController {
     private final ItemService itemService;
 
     @GetMapping("/{id}")
-    public ItemDto getItem(@PathVariable Integer id) {
+    public ItemDto getItem(@PathVariable Long id) {
         return itemService.getItem(id);
     }
 
     @GetMapping
-    public List<ItemDto> getAllItem(@RequestHeader("X-Sharer-User-Id") Integer id) {
+    public List<ItemDto> getAllItem(@RequestHeader("X-Sharer-User-Id") Long id) {
         return itemService.getAll(id);
     }
 
     @PostMapping
-    public ItemDto addItem(@Valid @RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Integer id) {
+    public ItemDto addItem(@Valid @RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Long id) {
         return itemService.addItem(itemDto, id);
     }
 
     @PatchMapping("/{id}")
-    public ItemDto updateItem(@RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Integer ownerId,
-                              @PathVariable("id") Integer itemId) {
+    public ItemDto updateItem(@RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Long ownerId,
+                              @PathVariable("id") Long itemId) {
         return itemService.updateItem(itemDto, ownerId, itemId);
     }
