+++ src/main/java/ru/practicum/shareit/item/storage/ItemStorage.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.item.model.Item;
 import java.util.List;
 
 public interface ItemStorage {
-    Item getItem(Integer itemId);
+    Item getItem(Long itemId);
 
     Item createItem(Item item);
 
     Item updateItem(Item item);
 
-    List<Item> getAllUserItem(Integer ownerId);
+    List<Item> getAllUserItem(Long ownerId);
 
     List<Item> searchItem(String text);
 }
